blog_management/requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
drf-yasg
django-filter
psycopg2-binary
python-decouple